+++ incubator/binding-risingwave/src/main/java/io/aklivity/zilla/runtime/binding/risingwave/internal/stream/RisingwaveProxyFactory.java
@@ -314,6 +314,10 @@ public final class RisingwaveProxyFactory implements RisingwaveStreamFactory
                 final DataFW data = dataRO.wrap(buffer, index, index + length);
                 onAppData(data);
                 break;
+            case FlushFW.TYPE_ID:
+                final FlushFW flush = flushRO.wrap(buffer, index, index + length);
+                onAppFlush(flush);
+                break;
             case EndFW.TYPE_ID:
                 final EndFW end = endRO.wrap(buffer, index, index + length);
                 onAppEnd(end);
@@ -410,6 +414,15 @@ public final class RisingwaveProxyFactory implements RisingwaveStreamFactory
                     doParseQuery(traceId, authorization);
                 }
             }
+        }
+
+        private void onAppFlush(
+            final FlushFW flush)
+        {
+            final long traceId = flush.traceId();
+            final long authorization = flush.authorization();
+
+            doAppWindow(traceId, authorization);
         }
 
         private void onAppEnd(